+++ lib/classy_enum/base.rb
@@ -8,7 +8,7 @@ module ClassyEnum
     include Translation
     include Collection
 
-    attr_accessor :owner, :serialize_as_json, :allow_blank
+    attr_accessor :owner, :allow_blank
 
     def base_class
       self.class.base_class
@@ -24,8 +24,6 @@ module ClassyEnum
       end
 
       def inherited(klass)
-        return if klass.anonymous?
-
         if self == ClassyEnum::Base
           klass.base_class = klass
         else
@@ -33,32 +31,6 @@ module ClassyEnum
           # Ensure subclasses follow expected naming conventions
           unless klass.name.start_with? "#{base_class.name}::"
             raise SubclassNameError, "subclass must be namespaced with #{base_class.name}::"
-          end
-
-          # Add visit_EnumMember methods to support validates_uniqueness_of with enum field
-          # This is due to a bug in Rails where it uses the method result as opposed to the
-          # database value for validation scopes. A fix will be released in Rails 4, but
-          # this will remain until Rails 3.x is no longer prevalent.
-          if defined?(Arel::Visitors::ToSql)
-            visitor_method = "visit_#{klass.name.split('::').join('_')}"
-
-            Arel::Visitors::ToSql.class_eval do
-              define_method visitor_method, lambda {|*values|
-                values[0] = values[0].to_s
-                begin
-                  quoted(*values)
-                rescue NoMethodError
-                  quote(*values)
-                end
-              }
-            end
-
-            Arel::Visitors::DepthFirst.class_eval do
-              define_method visitor_method, lambda {|*values|
-                values[0] = values[0].to_s
-                terminal(*values)
-              }
-            end
           end
 
           # Convert from MyEnumClass::NumberTwo to :number_two
@@ -88,18 +60,12 @@ module ClassyEnum
       def build(value, options={})
         object = find(value)
 
-        if object.nil? || (options[:allow_blank] && object.nil?)
-          if value.blank?
-            object = build_null_object(value)
-          else
-            return value
-          end
+        if object.nil?
+          value
+        else
+          object.owner = options[:owner]
+          object
         end
-
-        object.owner = options[:owner]
-        object.serialize_as_json = options[:serialize_as_json]
-        object.allow_blank = options[:allow_blank]
-        object
       end
 
       # DSL setter method for overriding reference to enum owner (ActiveRecord model)
@@ -116,20 +82,8 @@ module ClassyEnum
       #    end
       #  end
       def owner(owner)
-        define_method owner, lambda { @owner }
-      end
-
-      private
-
-      # Subclass the base class and make it behave like the value that it is
-      def build_null_object(value)
-        Class.new(base_class) {
-          @option = value
-          @index = 0
-          delegate :blank?, :nil?, :to => :option
-        }.new
+        define_method owner, -> { @owner }
       end
     end
-
   end
 end