+++ lib/generators/test_unit/classy_enum_generator.rb
@@ -3,8 +3,8 @@ module TestUnit
     class ClassyEnumGenerator < Rails::Generators::NamedBase
       desc "Generate a ClassyEnum test in test/enums/"
 
-      argument :name, :type => :string, :required => true, :banner => 'EnumName'
-      argument :values, :type => :array, :default => [], :banner => 'value1 value2 value3 etc...'
+      argument :name, type: :string, required: true, banner: 'EnumName'
+      argument :values, type: :array, default: [], banner: 'value1 value2 value3 etc...'
 
       source_root File.expand_path("../templates", __FILE__)