+++ spec/spec_helper.rb
@@ -5,7 +5,9 @@ require 'rubygems'
 require 'active_record'
 require 'classy_enum'
 
-ActiveRecord::Base.establish_connection(:adapter => "sqlite3", :database => ":memory:")
+I18n.enforce_available_locales = true
+
+ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
 
 RSpec.configure do |config|
   config.color_enabled = true