+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # ClassyEnum Changelog
 
+## 4.0.0 (Unreleased)
+
+* [BREAKING] Removes support for Ruby 1.8.7
+* [BREAKING] Removes support for Rails 3.0.x and 3.1.x
+* [BREAKING] Active Record models must explicitly include ClassyEnum::ActiveRecord now
+* [BREAKING] Removed use of null objects. Blank values are now returned as is from Enum.build.
+* [BREAKING] Removed serialize_as_json option. #as_json should be overriden in ClassyEnum::Base subclasses instead.
+* [BREAKING] Removed allow_blank option from Enum.build. This was used internally for legacy reasons and is no longer needed.
+* Prefer 'class_name' over 'enum' as optional class name argument
+
 ## 3.5.0
 
 * Fixes long standing issue with default values not being persisted to